move fns into HmsCatalog
