+++ crates/catalog/hms/src/catalog.rs
@@ -17,6 +17,8 @@
 
 use super::utils::*;
 use async_trait::async_trait;
+use hive_metastore::Database;
+use hive_metastore::PrincipalType;
 use hive_metastore::ThriftHiveMetastoreClient;
 use hive_metastore::ThriftHiveMetastoreClientBuilder;
 use hive_metastore::ThriftHiveMetastoreGetDatabaseException;
@@ -25,11 +27,18 @@ use iceberg::{
     Catalog, Error, ErrorKind, Namespace, NamespaceIdent, Result, TableCommit, TableCreation,
     TableIdent,
 };
+use pilota::AHashMap;
+use pilota::FastStr;
 use std::collections::HashMap;
 use std::fmt::{Debug, Formatter};
 use std::net::ToSocketAddrs;
 use typed_builder::TypedBuilder;
 use volo_thrift::ResponseError;
+
+/// hive.metastore.database.owner setting
+pub const HMS_DB_OWNER: &str = "hive.metastore.database.owner";
+/// hive.metastore.database.owner-type setting
+pub const HMS_DB_OWNER_TYPE: &str = "hive.metastore.database.owner-type";
 
 /// Which variant of the thrift transport to communicate with HMS
 /// See: <https://github.com/apache/thrift/blob/master/doc/specs/thrift-rpc.md#framed-vs-unframed-transport>
@@ -97,6 +106,154 @@ impl HmsCatalog {
             client: HmsClient(client),
         })
     }
+
+    /// Create and extract properties from `hive_metastore::hms::Database`.
+    pub fn properties_from_database(database: &Database) -> HashMap<String, String> {
+        let mut properties = HashMap::new();
+
+        if let Some(description) = &database.description {
+            properties.insert("comment".to_string(), description.to_string());
+        };
+
+        if let Some(location) = &database.location_uri {
+            properties.insert("location".to_string(), location.to_string());
+        };
+
+        if let Some(owner) = &database.owner_name {
+            properties.insert(HMS_DB_OWNER.to_string(), owner.to_string());
+        };
+
+        if let Some(owner_type) = &database.owner_type {
+            let value = match owner_type {
+                PrincipalType::User => "User",
+                PrincipalType::Group => "Group",
+                PrincipalType::Role => "Role",
+            };
+
+            properties.insert(HMS_DB_OWNER_TYPE.to_string(), value.to_string());
+        };
+
+        if let Some(params) = &database.parameters {
+            params.iter().for_each(|(k, v)| {
+                properties.insert(k.clone().into(), v.clone().into());
+            });
+        };
+
+        properties
+    }
+
+    /// Converts name and properties into `hive_metastore::hms::Database`
+    /// after validating the `namespace` and `owner-settings`.
+    pub fn convert_to_database(
+        namespace: &NamespaceIdent,
+        properties: &HashMap<String, String>,
+    ) -> Result<Database> {
+        let name = HmsCatalog::validate_namespace(namespace)?;
+        HmsCatalog::validate_owner_settings(properties)?;
+
+        let mut db = Database::default();
+        let mut parameters = AHashMap::new();
+
+        db.name = Some(name.into());
+
+        for (k, v) in properties {
+            match k.as_str() {
+                "comment" => db.description = Some(v.clone().into()),
+                "location" => {
+                    db.location_uri = Some(HmsCatalog::format_location_uri(v.clone()).into())
+                }
+                HMS_DB_OWNER => db.owner_name = Some(v.clone().into()),
+                HMS_DB_OWNER_TYPE => {
+                    let owner_type = match v.to_lowercase().as_str() {
+                        "user" => PrincipalType::User,
+                        "group" => PrincipalType::Group,
+                        "role" => PrincipalType::Role,
+                        _ => {
+                            return Err(Error::new(
+                                ErrorKind::DataInvalid,
+                                format!("Invalid value for setting 'owner_type': {}", v),
+                            ))
+                        }
+                    };
+                    db.owner_type = Some(owner_type);
+                }
+                _ => {
+                    parameters.insert(
+                        FastStr::from_string(k.clone()),
+                        FastStr::from_string(v.clone()),
+                    );
+                }
+            }
+        }
+
+        db.parameters = Some(parameters);
+
+        // Set default user, if none provided
+        // https://github.com/apache/iceberg/blob/main/hive-metastore/src/main/java/org/apache/iceberg/hive/HiveHadoopUtil.java#L44
+        if db.owner_name.is_none() {
+            db.owner_name = Some("user.name".into());
+            db.owner_type = Some(PrincipalType::User);
+        }
+
+        Ok(db)
+    }
+
+    /// Checks if provided `NamespaceIdent` is valid.
+    pub fn validate_namespace(namespace: &NamespaceIdent) -> Result<String> {
+        let name = namespace.as_ref();
+
+        if name.len() != 1 {
+            return Err(Error::new(
+                ErrorKind::DataInvalid,
+                "Invalid database, hierarchical namespaces are not supported",
+            ));
+        }
+
+        let name = name[0].clone();
+
+        if name.is_empty() {
+            return Err(Error::new(
+                ErrorKind::DataInvalid,
+                "Invalid database, provided namespace is empty.",
+            ));
+        }
+
+        Ok(name)
+    }
+
+    /// Formats location_uri by e.g. removing trailing slashes.
+    fn format_location_uri(location: String) -> String {
+        let mut location = location;
+
+        if !location.starts_with('/') {
+            location = format!("/{}", location);
+        }
+
+        if location.ends_with('/') && location.len() > 1 {
+            location.pop();
+        }
+
+        location
+    }
+
+    /// Checks if `owner-settings` are valid.
+    /// If `owner_type` is set, then `owner` must also be set.
+    fn validate_owner_settings(properties: &HashMap<String, String>) -> Result<()> {
+        let owner_is_set = properties.get(HMS_DB_OWNER).is_some();
+        let owner_type_is_set = properties.get(HMS_DB_OWNER_TYPE).is_some();
+
+        if owner_type_is_set && !owner_is_set {
+            return Err(Error::new(
+                ErrorKind::DataInvalid,
+                format!(
+                    "Setting '{}' without setting '{}' is not allowed",
+                    HMS_DB_OWNER_TYPE, HMS_DB_OWNER
+                ),
+            ));
+        }
+
+        Ok(())
+    }
 }
 
 #[async_trait]
@@ -146,7 +303,7 @@ impl Catalog for HmsCatalog {
         namespace: &NamespaceIdent,
         properties: HashMap<String, String>,
     ) -> Result<Namespace> {
-        let database = convert_to_database(namespace, &properties)?;
+        let database = HmsCatalog::convert_to_database(namespace, &properties)?;
 
         self.client
             .0
@@ -168,7 +325,7 @@ impl Catalog for HmsCatalog {
     /// - If there is an error querying the database, returned by
     /// `from_thrift_error`.
     async fn get_namespace(&self, namespace: &NamespaceIdent) -> Result<Namespace> {
-        let name = validate_namespace(namespace)?;
+        let name = HmsCatalog::validate_namespace(namespace)?;
 
         let db = self
             .client
@@ -177,7 +334,7 @@ impl Catalog for HmsCatalog {
             .await
             .map_err(from_thrift_error)?;
 
-        let properties = properties_from_database(&db);
+        let properties = HmsCatalog::properties_from_database(&db);
         let ns = Namespace::with_properties(NamespaceIdent::new(name.into()), properties);
 
         Ok(ns)
@@ -196,7 +353,7 @@ impl Catalog for HmsCatalog {
     /// - `Err(...)` if an error occurs during validation or the Hive Metastore
     /// query, with the error encapsulating the issue.
     async fn namespace_exists(&self, namespace: &NamespaceIdent) -> Result<bool> {
-        let name = validate_namespace(namespace)?;
+        let name = HmsCatalog::validate_namespace(namespace)?;
 
         let resp = self.client.0.get_database(name.clone().into()).await;
 
@@ -230,7 +387,7 @@ impl Catalog for HmsCatalog {
         namespace: &NamespaceIdent,
         properties: HashMap<String, String>,
     ) -> Result<()> {
-        let db = convert_to_database(namespace, &properties)?;
+        let db = HmsCatalog::convert_to_database(namespace, &properties)?;
 
         let name = match &db.name {
             Some(name) => name,
@@ -259,7 +416,7 @@ impl Catalog for HmsCatalog {
     /// - `Err(...)` signifies failure to drop the namespace due to validation  
     /// errors, connectivity issues, or Hive Metastore constraints.
     async fn drop_namespace(&self, namespace: &NamespaceIdent) -> Result<()> {
-        let name = validate_namespace(namespace)?;
+        let name = HmsCatalog::validate_namespace(namespace)?;
 
         self.client
             .0
@@ -300,5 +457,114 @@ impl Catalog for HmsCatalog {
 
     async fn update_table(&self, _commit: TableCommit) -> Result<Table> {
         todo!()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use iceberg::{Namespace, NamespaceIdent};
+
+    use super::*;
+
+    #[test]
+    fn test_properties_from_database() -> Result<()> {
+        let ns = NamespaceIdent::new("my_namespace".into());
+        let properties = HashMap::from([
+            ("comment".to_string(), "my_description".to_string()),
+            ("location".to_string(), "/my_location".to_string()),
+            (HMS_DB_OWNER.to_string(), "apache".to_string()),
+            (HMS_DB_OWNER_TYPE.to_string(), "User".to_string()),
+            ("key1".to_string(), "value1".to_string()),
+        ]);
+
+        let db = HmsCatalog::convert_to_database(&ns, &properties)?;
+
+        let expected = HmsCatalog::properties_from_database(&db);
+
+        assert_eq!(expected, properties);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_validate_owner_settings() {
+        let valid = HashMap::from([
+            (HMS_DB_OWNER.to_string(), "apache".to_string()),
+            (HMS_DB_OWNER_TYPE.to_string(), "user".to_string()),
+        ]);
+        let invalid = HashMap::from([(HMS_DB_OWNER_TYPE.to_string(), "user".to_string())]);
+
+        assert!(HmsCatalog::validate_owner_settings(&valid).is_ok());
+        assert!(HmsCatalog::validate_owner_settings(&invalid).is_err());
+    }
+
+    #[test]
+    fn test_convert_to_database() -> Result<()> {
+        let ns = NamespaceIdent::new("my_namespace".into());
+        let properties = HashMap::from([
+            ("comment".to_string(), "my_description".to_string()),
+            ("location".to_string(), "my_location".to_string()),
+            (HMS_DB_OWNER.to_string(), "apache".to_string()),
+            (HMS_DB_OWNER_TYPE.to_string(), "user".to_string()),
+            ("key1".to_string(), "value1".to_string()),
+        ]);
+
+        let db = HmsCatalog::convert_to_database(&ns, &properties)?;
+
+        assert_eq!(db.name, Some(FastStr::from("my_namespace")));
+        assert_eq!(db.description, Some(FastStr::from("my_description")));
+        assert_eq!(db.owner_name, Some(FastStr::from("apache")));
+        assert_eq!(db.owner_type, Some(PrincipalType::User));
+
+        if let Some(params) = db.parameters {
+            assert_eq!(params.get("key1".into()), Some(&FastStr::from("value1")));
+        }
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_convert_to_database_with_default_user() -> Result<()> {
+        let ns = NamespaceIdent::new("my_namespace".into());
+        let properties = HashMap::new();
+
+        let db = HmsCatalog::convert_to_database(&ns, &properties)?;
+
+        assert_eq!(db.name, Some(FastStr::from("my_namespace")));
+        assert_eq!(db.owner_name, Some(FastStr::from("user.name")));
+        assert_eq!(db.owner_type, Some(PrincipalType::User));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_validate_namespace() {
+        let valid_ns = Namespace::new(NamespaceIdent::new("ns".to_string()));
+        let empty_ns = Namespace::new(NamespaceIdent::new("".to_string()));
+        let hierarchical_ns = Namespace::new(
+            NamespaceIdent::from_vec(vec!["level1".to_string(), "level2".to_string()]).unwrap(),
+        );
+
+        let valid = HmsCatalog::validate_namespace(valid_ns.name());
+        let empty = HmsCatalog::validate_namespace(empty_ns.name());
+        let hierarchical = HmsCatalog::validate_namespace(hierarchical_ns.name());
+
+        assert!(valid.is_ok());
+        assert!(empty.is_err());
+        assert!(hierarchical.is_err());
+    }
+
+    #[test]
+    fn test_format_location_uri() {
+        let inputs = vec!["iceberg", "is/", "/nice/", "really/nice/", "/"];
+        let outputs = vec!["/iceberg", "/is", "/nice", "/really/nice", "/"];
+
+        inputs
+            .into_iter()
+            .zip(outputs.into_iter())
+            .for_each(|(inp, out)| {
+                let location = HmsCatalog::format_location_uri(inp.to_string());
+                assert_eq!(location, out);
+            })
     }
 }
